// src/modules/profissionais/enums/profissionais.enum.ts
export enum PronomeEnum {
  DR = 'DR',
  DRA = 'DRA',
  SR = 'SR',
  SRA = 'SRA',
  MASCULINO = 'MASCULINO',
  FEMININO = 'FEMININO',
}

export enum SexoEnum {
  MASCULINO = 'MASCULINO',
  FEMININO = 'FEMININO',
}

export enum TipoContratacaoEnum {
  CLT = 'CLT',
  PJ = 'PJ',
  AUTONOMO = 'AUTONOMO',
}

export enum TipoProfissionalEnum {
  REALIZANTE = 'REALIZANTE',
  SOLICITANTE = 'SOLICITANTE',
  AMBOS = 'AMBOS',
}

export enum TipoDocumentoProfissionalEnum {
  DIPLOMA = 'DIPLOMA',
  CERTIFICADOS = 'CERTIFICADOS',
  CRM = 'CRM',
  COMPROVANTE_RESIDENCIA = 'COMPROVANTE_RESIDENCIA',
  COMPROVANTE_BANCARIO = 'COMPROVANTE_BANCARIO',
  ULTRASSONOGRAFIA_GERAL = 'ULTRASSONOGRAFIA_GERAL',
  ULTRASSONOGRAFIA_DOPPLER = 'ULTRASSONOGRAFIA_DOPPLER',
  RADIOLOGIA = 'RADIOLOGIA',
  TOMOGRAFIA = 'TOMOGRAFIA',
}

export enum StatusDocumentoEnum {
  PENDENTE = 'PENDENTE',
  APROVADO = 'APROVADO',
  REPROVADO = 'REPROVADO',
  VENCIDO = 'VENCIDO',
}

export enum EstadoConselhoEnum {
  AC = 'AC',
  AL = 'AL',
  AP = 'AP',
  AM = 'AM',
  BA = 'BA',
  CE = 'CE',
  DF = 'DF',
  ES = 'ES',
  GO = 'GO',
  MA = 'MA',
  MT = 'MT',
  MS = 'MS',
  MG = 'MG',
  PA = 'PA',
  PB = 'PB',
  PR = 'PR',
  PE = 'PE',
  PI = 'PI',
  RJ = 'RJ',
  RN = 'RN',
  RS = 'RS',
  RO = 'RO',
  RR = 'RR',
  SC = 'SC',
  SP = 'SP',
  SE = 'SE',
  TO = 'TO',
}